Added support for device and fixed issue with the sync

+++ appachhisdk/src/main/java/com/appachhi/sdk/database/entity/FpsEntity.java
@@ -15,23 +15,23 @@ import com.google.gson.annotations.SerializedName;
         indices = {@Index(name = "fps_session_index", value = "session_id")})
 public class FpsEntity extends BaseEntity {
     @SerializedName("fps")
-    private double fps;
+    private int fps;
 
     @Ignore
     public FpsEntity(double fps, String sessionId, long sessionTime) {
         super(sessionId, sessionTime);
-        this.fps = fps;
+        this.fps = (int) fps;
     }
 
     public FpsEntity() {
     }
 
 
-    public double getFps() {
+    public int getFps() {
         return fps;
     }
 
-    public void setFps(double fps) {
+    public void setFps(int fps) {
         this.fps = fps;
     }
